initialised project structure
Initialised a .NET 9.0 Console App with the following new directories: BusinessLogic, Data, Domain and Tests.

In Domain: we will have all domain model classes
In BusinessLogic: we have the entire business logic, specifically the holiday search implementation.
In Data: this directory will hold the JSON files which we will load into the console app.
In Tests: we have this directory for unit tests