Amended HolidaySearch for List of HolidayPackages instead of just one, Added a new unit test for second search criteria input, added a loadHotels fn similair to LoadFlights
For the holiday search; implementation filters flights and hotels based on a criteria; then using linq query we create a holiday package for every combination of a matching flight and hotel.

Then in the holiday search packages are ordered by their total price; the package of lowest price comes first.

This new test checks that when "Any London Airport" is selected must be mapped to the correct list of London based airports and verifies if the flight's departure airport is one of the allowed london airports.